+++ core/_bootstrap.js
@@ -13,15 +13,20 @@
 		CKEDITOR.env.hc = false;
 	else {
 		// Check whether high contrast is active by creating a colored border.
-		var hcDetect = CKEDITOR.dom.element.createFromHtml( '<div style="width:0px;height:0px;position:absolute;left:-10000px;' +
-			'border: 1px solid;border-color: red blue;"></div>', CKEDITOR.document );
+		var hcDetect = CKEDITOR.dom.element.createFromHtml( '<div style="width:0;height:0;position:absolute;left:-10000px;' +
+			'border:1px solid;border-color:red blue"></div>', CKEDITOR.document );
 
 		hcDetect.appendTo( CKEDITOR.document.getHead() );
 
 		// Update CKEDITOR.env.
 		// Catch exception needed sometimes for FF. (#4230)
 		try {
-			CKEDITOR.env.hc = hcDetect.getComputedStyle( 'border-top-color' ) == hcDetect.getComputedStyle( 'border-right-color' );
+			var top = hcDetect.getComputedStyle( 'border-top-color' ),
+				right = hcDetect.getComputedStyle( 'border-right-color' );
+
+			// We need to check if getComputedStyle returned any value, because on FF
+			// it returnes empty string if CKEditor is loaded in hidden iframe. (#11121)
+			CKEDITOR.env.hc = !!( top && top == right );
 		} catch ( e ) {
 			CKEDITOR.env.hc = false;
 		}